+++ src/core/include/trogdor/lua/api/entities/luathing.h
@@ -115,34 +115,6 @@ namespace trogdor::entity {
                Place (or nil if none)
          */
          static int getLocation(lua_State *L);
-
-         /*
-            Wraps around Thing::setLocation(). If the Thing is also an Object
-            and has an owner, it will be removed from the owner's inventory
-            first.
-
-            DO NOT pass nil to remove this Thing from its location. For that,
-            use clearLocation instead.
-
-            Input:
-               Place (in Lua)
-
-            Output:
-               (none)
-         */
-         static int setLocation(lua_State *L);
-
-         /*
-            Removes Thing from whatever Place it's in. If it isn't in a place,
-            this function will do nothing.
-
-            Input:
-               (none)
-
-            Output:
-               (none)
-         */
-         static int removeFromLocation(lua_State *L);
    };
 }